feat: add new subcomponents
