+++ src/App.jsx
@@ -2,14 +2,12 @@
 import React, { useEffect, useMemo, useState, Suspense, lazy } from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 import { Card, CardHeader, CardContent, CardTitle } from "./components/ui/card";
-import { Button } from "./components/ui/button";
-import { Input } from "./components/ui/input";
-import { Loader2, Github, RefreshCcw, ShieldQuestion, Sun, Moon, Monitor, Maximize2, Minimize2, ExternalLink } from "lucide-react";
+import { Loader2, ShieldQuestion } from "lucide-react";
 import { githubGraphQL, ORG_REPOS_ISSUES, fetchProjectsWithStatus, fetchIssueTypes } from "./api/github";
 import useAppStore from "./store";
+import AppHeader from "./components/AppHeader";
 
 // Import components
-const Navigation = lazy(() => import("./components/Navigation"));
 const Dashboard = lazy(() => import("./components/Dashboard"));
 const ByAssignee = lazy(() => import("./components/ByAssignee"));
 const ByTags = lazy(() => import("./components/ByTags"));
@@ -269,42 +267,20 @@ export default function App() {
   }, [issueTypes]);
 
   return (
-    <div className="min-h-screen bg-gray-50">
-      <header className="sticky top-0 z-10 bg-white/80 backdrop-blur shadow-sm">
-        <div className="max-w-7xl mx-auto px-4 py-4">
-          <div className="flex items-center gap-3 mb-4">
-            <Github className="w-6 h-6" />
-            <h1 className="text-xl font-semibold">GitHub Issues Manager</h1>
-            <a
-              href="https://github.com/evolvus/github-issue-manager"
-              target="_blank"
-              rel="noopener noreferrer"
-              className="text-sm text-blue-600 hover:underline flex items-center gap-1"
-              title="View Source Code"
-            >
-              <Github className="w-4 h-4" />
-            </a>
-            <div className="ml-auto flex items-center gap-2 w-full sm:w-auto">
-              <Input placeholder="Organization (e.g. vercel)" value={org} onChange={e=>setOrg(e.target.value)} className="w-44" />
-              <Input placeholder="Personal Access Token" type="password" autoComplete="off" value={token} onChange={e=>setToken(e.target.value)} className="w-64" />
-              <Button className="bg-black text-white" onClick={loadData} disabled={loading || !org || !token}>
-                {loading ? (<><Loader2 className="w-4 h-4 animate-spin"/><span>Loading</span></>) : (<><RefreshCcw className="w-4 h-4"/><span>Load</span></>)}
-              </Button>
-              <button onClick={cycleTheme} title={`Theme: ${theme}`} className="p-2 border rounded">
-                {theme === "dark" ? <Moon className="w-4 h-4"/> : theme === "light" ? <Sun className="w-4 h-4"/> : <Monitor className="w-4 h-4"/>}
-              </button>
-              <button onClick={toggleDensity} title={`Density: ${density}`} className="p-2 border rounded">
-                {density === "compact" ? <Minimize2 className="w-4 h-4"/> : <Maximize2 className="w-4 h-4"/>}
-              </button>
-            </div>
-          </div>
-          {sidebarOpen && (
-            <Suspense fallback={<Loader2 className="w-4 h-4 animate-spin" />}>
-              <Navigation />
-            </Suspense>
-          )}
-        </div>
-      </header>
+      <div className="min-h-screen bg-gray-50">
+        <AppHeader
+          org={org}
+          setOrg={setOrg}
+          token={token}
+          setToken={setToken}
+          loadData={loadData}
+          loading={loading}
+          cycleTheme={cycleTheme}
+          toggleDensity={toggleDensity}
+          theme={theme}
+          density={density}
+          sidebarOpen={sidebarOpen}
+        />
 
       <main className="max-w-7xl mx-auto px-4 py-6">
         {error && (
